DocInfo transform must not use "rawsource" attribute for escaping.
Remove implementation of escaping author-separators in bibliographic fields
that relies on the "rawsource" attribute.
This is not safe (rawsource is only for information and debugging purposes).

A proper fix can be done with null-escaped text in the doctree.
C.f. https://sourceforge.net/p/docutils/bugs/_discuss/thread/c8f86be6/74ed/attachment/null-escape-in-doctree2.patch